Conditional reference assemblies from local folder
Game assemblies have been stripped and publicized in a local folder.
These changes conditionally use a hintpath if this folder is present.
It's a little hacky, and preferably would be moved to a common props
file instead of modifying each project. This change is a pre-cursor to
having the game assemblies accessible via a NuGet package.